+++ eppy/__init__.py
@@ -15,30 +15,41 @@ from __future__ import unicode_literals
 
 __author__ = """Santosh Philip"""
 __email__ = "santosh@noemail.com"
-__version__ = "0.5.63"
+__version__ = "0.5.66"
 
 
 from io import StringIO
 import eppy
+from eppy import modeleditor
+from eppy.modeleditor import IDF
 
 
 def newidf(version=None):
     """open a new idf file
 
-    easy way to open a new idf file for particular version. Works only id Energyplus of that version is installed.
+    easy way to open a new idf file for particular version. Works only if Energyplus of that version is installed.
+
+    - newidf(version=None) or newidf() will workif the IDD has already been set
+    - newidf(version=some_version) will work
+        - if some_version matches the IDD already set
+        - OR if no IDD has been set
 
     Parameters
     ----------
     version: string
-        version of the new file you want to create. Will work only if this version of Energyplus has been installed.
+        version of the new file you want to create. Will work only if this version of Energyplus has been installed OR if it matches the IDD already set.
 
     Returns
     -------
     idf
        file of type eppy.modelmake.IDF
     """  # noqa: E501
+    import eppy.easyopen as easyopen
+
     if not version:
-        version = "8.9"
+        return IDF(StringIO(""))
+        # return easyopen.easyopen(StringIO(""))
+
     import eppy.easyopen as easyopen
 
     idfstring = "  Version,{};".format(str(version))