+++ tests/test_easyopen.py
@@ -22,8 +22,14 @@ from eppy.pytest_helpers import do_integration_tests
 import eppy.idd_helpers as idd_helpers
 
 
-from six import StringIO
-from importlib import reload
+from io import StringIO
+
+from tests.pytest_helpers import safeIDDreset
+
+
+def teardown_module(module):
+    """new IDD has been set in the module. Here you tear it down"""
+    safeIDDreset()
 
 
 def test_cleanupversion():
@@ -49,9 +55,7 @@ def test_easyopen_idfopen():
     txt, result = ("  Version,{};".format(ver), "{}".format(ver))
     fhandle1 = StringIO(txt)
     fhandle2 = StringIO(txt)
-    reload(eppy)
-    reload(modeleditor)
-    reload(easyopen)
+    safeIDDreset()
     idf1, idf2 = easyopen.easyopen(fhandle1), eppy.openidf(fhandle2)
     for idf in [idf1, idf2]:
         versions = idf.idfobjects["version"]
@@ -80,9 +84,7 @@ def test_easyopen_withidd():
     txt, result = ("  Version,{};".format(ver), "{}".format(ver))
     fhandle1 = StringIO(txt)
     fhandle2 = StringIO(txt)
-    reload(eppy)
-    reload(modeleditor)
-    reload(easyopen)
+    safeIDDreset()
     idf1, idf2 = (
         easyopen.easyopen(fhandle1, idd=iddfile),
         eppy.openidf(fhandle2, idd=iddfile),