+++ tests/test_simpleread.py
@@ -10,7 +10,14 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 from io import StringIO
+import eppy
 import eppy.simpleread as simpleread
+from tests.pytest_helpers import safeIDDreset
+
+
+def teardown_module(module):
+    """new IDD has been set in the module. Here you tear it down"""
+    safeIDDreset()
 
 
 def test_idf2txt():